homeduino2mqtt/requirements.txt
paho-mqtt==1.4.0
pyserial==3.4
pyserial-asyncio==0.4